[learning] add learn_reset command
